Refactor - DRY'd up the JWKS functionality for the tests
